refactor: 회원과 인증 관련 Controller 위치 이동
- 토큰 관련 로직이 있을 경우 AuthController 에 위치
- 토큰 관련 로직 없이 사용자 관련 로직만 있을 경우 CustomerController 에 위치
